refactor&feat: removed testify package and added accept header handling logic for both servers
